refactor: cleanup controller
